requirements.txt
opencv-python
foxglove_sdk
foxglove_websocket